+++ src/Services/IM/IMServiceBuilder.php
@@ -26,6 +26,7 @@ use Bitrix24\SDK\Services\IM\Placements\PlacementLocationCodes;
 use Bitrix24\SDK\Services\IM\Placements\Placements;
 use Bitrix24\SDK\Services\IM\Recent\Service\Recent;
 use Bitrix24\SDK\Services\IM\Revision\Service\Revision;
+use Bitrix24\SDK\Services\IM\Search\Service\Search;
 use Bitrix24\SDK\Services\IM\User\Service\UserStatus;
 use Bitrix24\SDK\Services\IM\User\Service\User;
 use Bitrix24\SDK\Services\Placement\Service\Placement;
@@ -37,6 +38,15 @@ class IMServiceBuilder extends AbstractServiceBuilder
     {
         if (!isset($this->serviceCache[__METHOD__])) {
             $this->serviceCache[__METHOD__] = new Recent($this->core, $this->log);
+        }
+
+        return $this->serviceCache[__METHOD__];
+    }
+
+    public function search(): Search
+    {
+        if (!isset($this->serviceCache[__METHOD__])) {
+            $this->serviceCache[__METHOD__] = new Search($this->core, $this->log);
         }
 
         return $this->serviceCache[__METHOD__];